requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.5
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
redis==6.2.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
psycopg==3.2.6
psycopg-binary==3.2.6
openpyxl>=3.1.2
xhtml2pdf
